Refactor OrganizationAPI into using a generic one using a namespace called EntityApi
